// src/app/models/user.model.ts
export class Users {
    id:string;
    username: string;
    password: string;
    confirmPassword: string;
    email: string;
    lastName: string;
    firstName: string;
    middleName: string;
    extensionName: string;
    fullName:string;
    dateOfBirth: string;
    designation: string;
    expirationDate: string;
    isActive: boolean;
    prefix: string;
    position: string;
    mobileNo: string;
    userTypeId: string;
    license: string;
    licenseExpiration: string;
    phicAccreditationNo: string;
    phicAccreditationExpiration: string;
    s2License: string;
    s2LicenseExpiration: string;
    appType: string;
    appId: string;
}